GeoLight-package {GeoLight}	R Documentation

The GeoLight Package

Description

This is a summary of all features of GeoLight, a R-package for analyzing light based geolocator data

Details

GeoLight is a package to derive geographical positions from daily light intensity pattern. Positioning and calibration methods are based on the threshold-method (Ekstrom 2004, Lisovski et al. 2012). A changepoint model from the R package changepoint is implemented to distinguish between periods of residency and movement based on the sunrise and sunset times. Mapping functions are implemented using the R package maps.

Getting Started

We refrain from giving detailed background on the (several steps of) analysis of light-based geolocator data here but strongly recommend the key-publications below.

Updates

We advise all users to update their installation of GeoLight regularly. Type news(package="GeoLight") to read news documentation about changes to the recent and all previous version of the package

Important notes

Most functions in GeoLight require the same initial units and mostly the format and object type is mandatory:

tFirst 	 yyyy-mm-dd hh:mm "UTC" (see: as.POSIXct, time zones)

tSecond 	 as tFirst (e.g. 2008-12-01 17:30)

type 	 either 1 or 2 depending on wheter tFirst is sunrise (1) or sunset (2)

coord 	 SpatialPoints or a matrix, containing x and y coordinates (in that order)

degElevation 	 a vector or a single value of sun elevation angle(s) in degrees (e.g. -6)

FUNCTIONS AND DATASETS

In the following, we give a summary of the main functions and sample datasets in the GeoLight package. Alternatively a list of all functions and datasets in alphabetical order is available by typing library(help=GeoLight). For further information on any of these functions, type help(function name).

CONTENTS

I. 	 Determination of sunset and sunrise

II. 	 Residency analysis

III. 	 Calibration

IV. 	 Positioning

V. 	 Data visualisation

VI. 	 Examples

I. Determination of sunset and sunrise

gleTrans 	 transformation of already defined twilight events*

glfTrans 	 transformation of light intensity measurements over time*

luxTrans 	 transformation of light intensity measurements over time**

lightFilter 	 filter to remove noise in light intensity measurements during the night

twilightCalc 	 definition of twilight events (sunrise, sunset) from light intensity measurements

* written for data recorded by geolocator devices from the Swiss Ornithological Institute
** written for data recorded by geolocator devices from Migrate Technology Ltd

II. Residency Analysis

changeLight 	 function to distinguish between residency and movement periods

schedule 	 function to produce a data frame summerizing the residency and movement pattern

III. Calibration

See Lisovski et al. 2012 for all implemented calibration methods.

getElevation 	 function to calculate the sun elevation angle for data with known position

HillEkstromCalib 	 Hill-Ekstrom calibration for one or more defined stationary periods

IV. Positioning

coord 	 main function to derive a matrix of spatial coordinates

distanceFilter 	 filter function to reduce unrealistic positions (not recommended, since the filtering ignore positioning error)

loessFilter 	 filter function to define outliers in sunrise and sunset times (defined twilight events)

V. Data visualisation

tripMap 	 function to map the derived positions and combine the coordinates in time order

siteMap 	 function to show the results of the residency analysis on a map

IV. Examples

calib1 	 data for calibration: light intensities

calib2 	 data for calibration: Calculated twilight events (from calib1 by twilightCalc)

hoopoe1 	 light intensity measurements over time recorded on a migratory bird

hoopoe2 	 sunrise and sunset times: From light intensity measurement (from hoopoe1)

R Packages for Further Spatial Ananlyses

spatstat
adehabitat
gstat
trip
tripEstimation
move
...

Acknowledgements

Steffen Hahn, Felix Liechti, Fraenzi Korner-Nievergelt, Andrea Koelzsch, Eldar Rakhimberdiev, Erich Baechler, Eli Bridge, Andrew Parnell, Richard Inger

Authors

Simeon Lisovski, Simon Wotherspoon, Michael Sumner, Silke Bauer, Tamara Emmenegger
Maintainer: Simeon Lisovski <simeon.lisovski(at)gmail.com>

References

Ekstrom, P.A. (2004) An advance in geolocation by light. Memoirs of the National Institute of Polar Research, Special Issue, 58, 210-226.

Fudickar, A.M., Wikelski, M., Partecke, J. (2011) Tracking migratory songbirds: accuracy of light-level loggers (geolocators) in forest habitats. Methods in Ecology and Evolution, DOI: 10.1111/j.2041-210X.2011.00136.x.

Hill, C. & Braun, M.J. (2001) Geolocation by light level - the next step: Latitude. Electronic Tagging and Tracking in Marine Fisheries (eds J.R. Sibert & J. Nielsen), pp. 315-330. Kluwer Academic Publishers, The Netherlands.

Hill, R.D. (1994) Theory of geolocation by light levels. Elephant Seals: Population Ecology, Behavior, and Physiology (eds L. Boeuf, J. Burney & R.M. Laws), pp. 228-237. University of California Press, Berkeley.

Lisovski, S. and Hahn, S. (2012) GeoLight - processing and analysing light-based geolocator data in R. Methods in Ecology and Evolution, doi: 10.1111/j.2041-210X.2012.00248.x

Lisovski, S., Hewson, C.M, Klaassen, R.H.G., Korner-Nievergelt, F., Kristensen, M.W & Hahn, S. (2012) Geolocation by light: Accuracy and precision affected by environmental factors. Methods in Ecology and Evolution, doi: 10.1111/j.2041-210X.2012.00185.x

Wilson, R.P., Ducamp, J.J., Rees, G., Culik, B.M. & Niekamp, K. (1992) Estimation of location: global coverage using light intensity. Wildlife telemetry: remote monitoring and tracking of animals (eds I.M. Priede & S.M. Swift), pp. 131-134. Ellis Horward, Chichester.

[Package GeoLight version 2.0.0 Index]