distanceFilter {GeoLight}	R Documentation

Filter for unrealistic positions within a track based on distance

Description

The filter identifies unrealistic positions. The maximal distance per hour/day can be set corresponding to the particular species.

Usage

distanceFilter(tFirst, tSecond, type, twl, degElevation = -6, distance,
  units = "hour")

Arguments

tFirst
vector of sunrise/sunset times (e.g. 2008-12-01 08:30).

tSecond
vector of of sunrise/sunset times (e.g. 2008-12-01 17:30).

type
vector of either 1 or 2, defining tFirst as sunrise or sunset respectively.

twl
data.frame containing twilights and at least tFirst, tSecond and type (alternatively give each parameter separately).

degElevation
sun elevation angle in degrees (e.g. -6 for "civil twilight")

distance
the maximal distance in km per units. Distances above will be considered as unrealistic.

units
the time unite corresponding to the distance. Default is "hour", alternative option is "day".

Details

Note that this type of filter significantly depends on the calibration (degElevation). Especially during equinox periods. In contrast, the (loessFilter) is independent from positions (uses twilight times) and therefore superior.

Value

Logical vector. TRUE means the particular position passed the filter.

Author(s)

Simeon Lisovski, Fraenzi Korner-Nievergelt

[Package GeoLight version 2.0.0 Index]