gleTrans {GeoLight}	R Documentation

Transformation of *.gle files

Description

Function to transform *.gle files derived by the software GeoLocator for further analyses in GeoLight.

Usage

gleTrans(file)

Arguments

file
the full patch and filename with suffix of the *.gle file.

Details

The *.gle file derived by the software "GeoLocator" (Swiss Ornithological Institute) is a table with interpolated light intensities over time gathered from the *.glf file. Furthermore a column defines wether the light intensity passes the defined light intensity threshold in the morning (sunrise) or in the evening (sunset). This information is used in gleTrans(), to create a table with two subsequent twilight events (tFirst, tSecond) and type defining wether tFirst refers to sunrise (1) or sunset (2). Date and time information will be transferred into a GeoLight appropriate format (see: as.POSIXct).

Value

A data.frame suitable for further use in GeoLight.

Author(s)

Simeon Lisovski

See Also

glfTrans

[Package GeoLight version 2.0.0 Index]