glfTrans {GeoLight}	R Documentation

Transformation of *.glf files

Description

Transform *.glf files derived by the software GeoLocator for further analyses in GeoLight.

Usage

glfTrans(file = "/path/file.glf")

Arguments

file
the full patch and filename with suffix of the *.glf file.

Details

The *.glf files produced by the software GeoLocator (Swiss Ornithological Institute) is a table with light intensity measurements over time. glfTrans produces a table with these measurements and transfer the data and time information into the format required by GeoLight format (see: as.POSIXct).

Value

A data.frame suitable for further use in GeoLight.

Author(s)

Simeon Lisovski

See Also

gleTrans; luxTrans for transforming *.lux files produced by Migrate Technology Ltd

[Package GeoLight version 2.0.0 Index]