hoopoe1 {GeoLight}	R Documentation

Light intensity measurements over time recorded on a migratory bird

Description

Sunlight intensity measurements over time recorded during the first part of the annual migration of a European Hoopoe (Upupa epops). All dates/times are measured in Universal Time Zone (UTC).

Format

A table with 24474 rows and 2 columns, rows corresponding to light measurements recorded in ten-minute intervals (datetime).

Source

Baechler, E., Hahn, S., Schaub, M., Arlettaz, R., Jenni, L., Fox, J.W., Afanasyev, V. & Liechti, F. (2010) Year-Round Tracking of Small Trans-Saharan Migrants Using Light-Level Geolocators. Plos One, 5.

[Package GeoLight version 2.0.0 Index]