ligTrans {GeoLight}	R Documentation

Transformation of *.lig files

Description

Transform *.lig files derived from geolocator deviced for further analyses in GeoLight.

Usage

ligTrans(file)

Arguments

file
the full patch and filename with suffix of the *.lux file.

Value

A data.frame suitable for further use in GeoLight.

Author(s)

Tamara Emmenegger

See Also

gleTrans for transforming *.glf files produced by the software GeoLocator (Swiss Ornithological Institute)

[Package GeoLight version 2.0.0 Index]