loessFilter {GeoLight}	R Documentation

Filter to remove outliers in defined twilight times based on smoother function

Description

This filter defines outliers based on residuals from a local polynomial regression fitting provcess (loess).

Usage

loessFilter(tFirst, tSecond, type, twl, k = 3, plot = TRUE)

Arguments

tFirst
vector of sunrise/sunset times (e.g. 2008-12-01 08:30).

tSecond
vector of of sunrise/sunset times (e.g. 2008-12-01 17:30).

type
vector of either 1 or 2, defining tFirst as sunrise or sunset respectively.

twl
data.frame containing twilights and at least tFirst, tSecond and type (alternatively give each parameter separately).

k
a measure of how many interquartile ranges to take before saying that a particular twilight event is an outlier

plot
codelogical, if TRUE a plot indicating the filtered times will be produced.

Value

Logical vector matching positions that pass the filter.

Author(s)

Simeon Lisovski & Eldar Rakhimberdiev

[Package GeoLight version 2.0.0 Index]