luxTrans {GeoLight}	R Documentation

Transformation of *.lux files

Description

Transform *.lux files derived from Migrate Technology Ltd geolocator deviced for further analyses in GeoLight.

Usage

luxTrans(file)

Arguments

file
the full patch and filename with suffix of the *.lux file.

Details

The *.lux files produced by Migrate Technology Ltd are table with light intensity measurements over time. luxTrans produces a table with these measurements and transfer the data and time information into the format required by GeoLight format (see: as.POSIXct).

Value

A data.frame suitable for further use in GeoLight.

Author(s)

Simeon Lisovski

See Also

gleTrans for transforming *.glf files produced by the software GeoLocator (Swiss Ornithological Institute)

[Package GeoLight version 2.0.0 Index]