refracted {GeoLight}	R Documentation

Atmospheric Refraction

Description

Adjust the solar zenith angle for atmospheric refraction.

Usage

refracted(zenith)

Arguments

zenith
zenith angle (degrees) to adjust.

Details

Given a vector of solar zeniths computed by zenith, refracted calculates the solar zeniths adjusted for the effect of atmospheric refraction.

unrefracted is the inverse of refracted. Given a (single) solar zenith adjusted for the effect of atmospheric refraction, unrefracted calculates the solar zenith as computed by zenith.

Value

vector of zenith angles (degrees) adjusted for atmospheric refraction.

Examples

## Refraction causes the sun to appears higher on the horizon
refracted(85:92)
## unrefracted gives unadjusted zenith (see SGAT)

[Package GeoLight version 2.0.0 Index]