schedule {GeoLight}	R Documentation

Function for making a data frame summarising residency and movement pattern.

Description

Function for making a data frame summarising residency and movement pattern.

Usage

schedule(tFirst, tSecond, site)

Arguments

tFirst
date and time of sunrise/sunset (e.g. 2008-12-01 08:30)

tSecond
date and time of sunrise/sunset (e.g. 2008-12-01 17:30)

site
a vector, indicating the residency period of a particular day (see output: changeLight)

Value

A data.frame with end and start date (yyyy-mm-dd hh:mm, UTC) for each stationary period.

Author(s)

Simeon Lisovski

Examples

data(hoopoe2)
  hoopoe2$tFirst <- as.POSIXct(hoopoe2$tFirst, tz = "GMT")
  hoopoe2$tSecond <- as.POSIXct(hoopoe2$tSecond, tz = "GMT")
residency <- changeLight(hoopoe2, rise.prob=0.1, set.prob=0.1, plot=FALSE, summary=FALSE)
schedule(hoopoe2[,1], hoopoe2[,2], site = residency$site)

[Package GeoLight version 2.0.0 Index]