solar {GeoLight}	R Documentation

Solar Time and Declination

Description

Calculate solar time, the equation of time and solar declination

Usage

solar(tm)

Arguments

tm
a vector of POSIXct times.

Details

The solar time, the equation of time and the sine and cosine of the solar declination are calculted for the times specified by tm using the same methods as www.esrl.noaa.gov/gmd/grad/solcalc/.

Value

A list containing the following vectors.

solarTime
the solar time (degrees)

eqnTime
the equation of time (minutes of time)

sinSolarDec
sine of the solar declination

cosSolarDec
cosine of the solar declination

See Also

zenith

Examples

## Current solar time
solar(Sys.time())

[Package GeoLight version 2.0.0 Index]