staroddiTrans {GeoLight}	R Documentation

Transformation of staroddi files

Description

Transform staroddi files derived from geolocator deviced for further analyses in GeoLight.

Usage

staroddiTrans(file)

Arguments

file
the full patch and filename of the staroddi file.

Value

A data.frame suitable for further use in GeoLight.

Author(s)

Tamara Emmenegger

[Package GeoLight version 2.0.0 Index]