tripMap {GeoLight}	R Documentation

Draw the positions and the trip on a map

Description

Draw a map (from the R Package maps) with calculated positions connected by a line

Usage

tripMap(crds, equinox = TRUE, map.range = c("EuroAfrica", "AustralAsia",
  "America", "World"), legend = TRUE, ...)

Arguments

crds
a SpatialPoints or matrix object, containing x and y coordinates (in that order).

equinox
logical; if TRUE, the equinox period(s) is shown as a broken blue line.

map.range
some possibilities to choose defined areas (default: "World").

legend
logical; if TRUE, a legend will be added to the plot.

...
Arguments to be passed to methods, such as graphical parameters (see par).

Author(s)

Simeon Lisovski

Examples

data(hoopoe2)
 hoopoe2$tFirst <- as.POSIXct(hoopoe2$tFirst, tz = "GMT")
 hoopoe2$tSecond <- as.POSIXct(hoopoe2$tSecond, tz = "GMT")
crds <- coord(hoopoe2, degElevation = -6)
tripMap(crds, xlim = c(-20,20), ylim = c(0,60), main="hoopoe2")

[Package GeoLight version 2.0.0 Index]