zenith {GeoLight}	R Documentation

Solar Zenith Angle

Description

Calculate the solar zenith angle for given times and locations

Usage

zenith(sun, lon, lat)

Arguments

sun
list of solar time and declination computed by solar.

lon
vector of longitudes.

lat
vector latitudes.

Details

zenith uses the solar time and declination calculated by solar to compute the solar zenith angle for given times and locations, using the same methods as www.esrl.noaa.gov/gmd/grad/solcalc/. This function does not adjust for atmospheric refraction see refracted.

Value

A vector of solar zenith angles (degrees) for the given locations and times.

See Also

solar

Examples

## Approx location of Sydney Harbour Bridge
lon <- 151.211
lat <- -33.852
## Solar zenith angle for noon on the first of May 2000
## at the Sydney Harbour Bridge
s <- solar(as.POSIXct("2000-05-01 12:00:00","EST"))
zenith(s,lon,lat)

[Package GeoLight version 2.0.0 Index]